+++ appyx-components/experimental/cards/android/build.gradle.kts
@@ -31,9 +31,9 @@ android {
 dependencies {
     val composeBom = platform(libs.compose.bom)
 
-    api(project(":appyx-interactions:android"))
     api(project(":demos:common"))
 
+    implementation(project(":appyx-interactions:android"))
     implementation(project(":appyx-interactions:appyx-interactions"))
     implementation(project(":appyx-components:experimental:cards:cards"))
     implementation(composeBom)
