+++ appyx-components/experimental/promoter/android/build.gradle.kts
@@ -32,7 +32,6 @@ dependencies {
     val composeBom = platform(libs.compose.bom)
 
     api(project(":appyx-interactions:android"))
-    api(project(":demos:common"))
 
     implementation(project(":appyx-interactions:appyx-interactions"))
     implementation(project(":appyx-components:experimental:promoter:promoter"))
