+++ appyx-components/internal/test-drive/android/src/androidTest/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.bumble.appyx.components.testdrive">
+    package="com.bumble.appyx.components.testdrive.android">
 
     <application android:theme="@android:style/Theme.Material.NoActionBar" />
