+++ appyx-components/internal/test-drive/android/src/main/kotlin/com/bumble/appyx/components/testdrive/android/TestDriveExperiment.kt
@@ -1,4 +1,4 @@
-package com.bumble.appyx.components.testdrive
+package com.bumble.appyx.components.testdrive.android
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxSize
